+++ backend/eodi-api/src/main/java/com/bb/eodi/deal/application/input/FindRecommendedSellInput.java
@@ -7,15 +7,15 @@ import java.util.List;
  */
 public record FindRecommendedSellInput(
         // 보유 현금
-        Integer cash,
+        Long cash,
         // 대상 지역 ID list
         List<Long> targetRegionIds,
         // 대상 주택 유형 list
         List<String> targetHousingTypes,
         // 최대 가격 (만원 단위)
-        Integer maxPrice,
+        Long maxPrice,
         // 최소 가격 (만원 단위)
-        Integer minPrice,
+        Long minPrice,
         // 최대전용면적
         Integer maxNetLeasableArea,
         // 최소전용면적
