+++ backend/eodi-api/src/main/java/com/bb/eodi/deal/presentation/dto/request/RealEstateSellRecommendRequestParameter.java
@@ -17,7 +17,7 @@ public record RealEstateSellRecommendRequestParameter(
         @Parameter(description = "보유 현금", example = "70000", required = true)
         @NotNull(message = "보유 현금은 필수값입니다.")
         @Positive(message = "보유 현금은 0보다 커야합니다.")
-        Integer cash,
+        Long cash,
 
         @Parameter(
                 description = "조회 대상 지역 ID 목록",
@@ -50,9 +50,9 @@ public record RealEstateSellRecommendRequestParameter(
          * ============ 필터 조건 ============
          */
         @Parameter(description = "최대 가격 (단위: 만원)", example = "34000")
-        Integer maxPrice,
+        Long maxPrice,
         @Parameter(description = "최소 가격 (단위: 만원)", example = "24000")
-        Integer minPrice,
+        Long minPrice,
         @Parameter(description = "최대 전용면적", example = "84")
         Integer maxNetLeasableArea,
         @Parameter(description = "최소 전용면적", example = "39")
